script: Move Castable/DerivedFrom/IDLInterface to script_bindings.
