+++ components/script_bindings/conversions.rs
@@ -19,9 +19,19 @@ use js::rust::{
 };
 use servo_config::opts;
 
+use crate::inheritance::Castable;
 use crate::reflector::Reflector;
 use crate::str::{ByteString, DOMString, USVString};
 use crate::utils::{DOMClass, DOMJSClass};
+
+/// A trait to check whether a given `JSObject` implements an IDL interface.
+pub trait IDLInterface {
+    /// Returns whether the given DOM class derives that interface.
+    fn derives(_: &'static DOMClass) -> bool;
+}
+
+/// A trait to mark an IDL interface as deriving from another one.
+pub trait DerivedFrom<T: Castable>: Castable {}
 
 // http://heycam.github.io/webidl/#es-USVString
 impl ToJSValConvertible for USVString {
