+++ components/script_bindings/script_runtime.rs
@@ -1,0 +1,17 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at https://mozilla.org/MPL/2.0/. */
+
+use std::cell::Cell;
+
+thread_local!(
+    static THREAD_ACTIVE: Cell<bool> = const { Cell::new(true) };
+);
+
+pub fn runtime_is_alive() -> bool {
+    THREAD_ACTIVE.with(|t| t.get())
+}
+
+pub fn mark_runtime_dead() {
+    THREAD_ACTIVE.with(|t| t.set(false));
+}
